Code Fixes for an Issue from BSE API
BseHelper.py : The issue with BSE API has been addressed temporarily.
Issue Description : While Calling GetAllBseScrips it should return a dataframe of all the BSE Stocks. However it has been observed that the INDUSTRY column is coming as NULL or Empty. Upon Further investigation is has been found that the BSE API https://api.bseindia.com/BseIndiaAPI/api/ListofScripData/w?Group=&Scripcode=&segment=EQT0&status=Active it self is giving the NULL values . So in such case We have tried to fill the INDUSTRY NAME from a older file 2025-10-08-22-43-21-NSE-BSE-IS-ALL-EQ.CSV. So we will continue use this strategy till the API from BSE end is fixed. once fixed the Data from BSE Api will be there so since the Values will not be NULL or empty we will not refer to the older file.